refactor(imports): consolidate import modules into parse/plan/execute seams
Merge the 15 import-* modules into five cohesive seams that form a clean
dependency DAG (parse -> plan -> execute -> sessions; presentation standalone):
- import-parse: session/JSONL parsing and branch selection
- import-plan: plan building plus checkpoint/manifest provenance
- import-execute: queue identity, delivery, noise/curation policy, retain payloads
- import-presentation: preview/result rendering (unchanged)
- import-sessions: execution + chat-transcript + operations orchestration and public API

Behavior, retain payloads, checkpoint/manifest format, and import identity are
unchanged (structure-only); the sessions<->execution cycle now lives within one
seam. Updates the code-map generator and all internal/test import paths.

Refs #416
