+++ tests/golden-memory-stability.test.ts
@@ -7,8 +7,8 @@ import { DEFAULT_CONFIG } from "../extensions/config/config.js";
 import { createMemoryOperations } from "../extensions/operations/memory-operation-service.js";
 import { createRetainTurnPolicy } from "../extensions/lifecycle/memory-lifecycle-retain.js";
 import type { RuntimeSnapshot } from "../extensions/lifecycle/memory-lifecycle-runtime.js";
-import { readImportCheckpoint } from "../extensions/imports/import-checkpoint.js";
-import { readImportManifest } from "../extensions/imports/import-manifest.js";
+import { readImportCheckpoint } from "../extensions/imports/import-plan.js";
+import { readImportManifest } from "../extensions/imports/import-plan.js";
 import { readRetainQueue, resolveQueuePath } from "../extensions/queue/queue.js";
 import { liveDocumentId, stableSessionId } from "../extensions/utils/session.js";
 import type { HindsightLikeClient, ResolvedConfig } from "../extensions/types.js";
